renamed urlmanager package to helpers, created new package and added new ProductShowActivity
