+++ app/src/main/java/com/example/cleanenvi/MainActivity.java
@@ -9,12 +9,8 @@ import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
-import android.widget.Toast;
 
-import com.example.cleanenvi.urlmanager.URLManager;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
-
-import java.io.IOException;
 
 public class MainActivity extends AppCompatActivity {
